requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
python-dotenv>=0.19.0
groq>=0.3.0
requests>=2.26.0
pillow>=9.0.0
numpy>=1.21.0
python-multipart>=0.0.5
langchain>=0.0.200
langchain-community>=0.0.20
ragas>=0.0.19
pypdf>=3.8.0
python-dateutil>=2.8.2
langchain_gorq
google-generativeai
crewai